better org of types
